Restructure model subpackages
