--- lume_model/keras/__init__.py
@@ -1,4 +1,0 @@
-try:
-    from lume_model.keras.model import KerasModel
-except ModuleNotFoundError:
-    pass
