--- lume_model/torch/__init__.py
@@ -1,5 +1,0 @@
-try:
-    from lume_model.torch.model import TorchModel
-    from lume_model.torch.module import TorchModule
-except ModuleNotFoundError:
-    pass
