+++ tests/models/test_keras_model.py
@@ -7,7 +7,7 @@ import pytest
 import numpy as np
 
 try:
-    from lume_model.keras import KerasModel
+    from lume_model.models import KerasModel
     from lume_model.variables import InputVariable, OutputVariable, ScalarOutputVariable
 except ImportError:
     pass
